USB: initial merge relatively complete, linking issues left

+++ pcsx2/USB/usb-pad/usb-pad.cpp
@@ -164,7 +164,7 @@ static void pad_handle_data(USBDevice *dev, USBPacket *p)
 		if (devep == 1 && s->pad) {
 			ret = s->pad->TokenIn(data, p->iov.size);
 			if (ret > 0)
-				usb_packet_copy (p, data, MIN(ret, sizeof(data)));
+				usb_packet_copy (p, data, MIN(ret, (int)sizeof(data)));
 			else
 				p->status = ret;
 		} else {
